.devcontainer/testing/requirements.txt
pytest-playwright
pytest-timeout
loguru